Move projections from methods and fix dot
