+++ pymrcpp/pyMethods.cpp
@@ -17,25 +17,10 @@
 #include "treebuilders/apply.h"
 //#include "treebuilders/grid.h"
 #include "treebuilders/project.h"
-#include "pyProject.h"
+//#include "pyProject.h"
 
 namespace py = pybind11;
 using namespace mrcpp;
-
-void pyProject1D(py::module &m) {
-    m.def("project", py::overload_cast<double, FunctionTree<1> &, std::function<double (double)>, int>(&project1D),
-    py::arg("prec"), py::arg("out"), py::arg("func"), py::arg("maxIter")= -1);
-}
-
-void pyProject2D(py::module &m) {
-    m.def("project", py::overload_cast<double, FunctionTree<2> &, std::function<double (double, double)>, int>(&project2D),
-    py::arg("prec"), py::arg("out"), py::arg("func"), py::arg("maxIter")= -1);
-}
-
-void pyProject3D(py::module &m) {
-    m.def("project", py::overload_cast<double, FunctionTree<3> &, std::function<double (double, double, double)>, int>(&project3D),
-    py::arg("prec"), py::arg("out"), py::arg("func"), py::arg("maxIter")= -1);
-}
 
 template <int D>
 void pyMethods(py::module &m) {
@@ -62,7 +47,7 @@ void pyMethods(py::module &m) {
         py::arg("prec"), py::arg("out"), py::arg("oper"), py::arg("inp"), py::arg("maxIter") = -1);
     m.def("apply", py::overload_cast<FunctionTree<D> &, DerivativeOperator<D> &, FunctionTree<D> &, int>(&apply<D>));
 
-//    m.def("dot", &dot<D>);
+    m.def("dot", py::overload_cast<FunctionTree<D> &, FunctionTree<D> &>(&dot<D>));
 }
 
 template void pyMethods<1>(py::module &m);
